+++ charts/latest/blobfuse-csi-driver/templates/csi-blobfuse-node.yaml
@@ -66,7 +66,7 @@ spec:
             requests:
               cpu: 10m
               memory: 20Mi
-        - name: blobfuse
+        - name: blob
           image: "{{ .Values.image.blobfuse.repository }}:{{ .Values.image.blobfuse.tag }}"
           args:
             - "--v=5"
@@ -117,7 +117,7 @@ spec:
               readOnly: true
               name: msi
             - mountPath: /mnt
-              name: blobfuse-cache
+              name: blob-cache
           resources:
             limits:
               cpu: 4
@@ -147,4 +147,4 @@ spec:
           name: msi
         - hostPath:
             path: /mnt
-          name: blobfuse-cache
+          name: blob-cache
