+++ pkg/blob/azure.go
