+++ pkg/blob/azure_test.go
