+++ pkg/blobplugin/Dockerfile
@@ -18,10 +18,10 @@ ADD . .
 ARG TARGETARCH
 ARG TARGETOS
 ARG LDFLAGS
-RUN CGO_ENABLED=0 GOOS=$TARGETOS GOARCH=$TARGETARCH go build -a -ldflags "${LDFLAGS:--X sigs.k8s.io/blobfuse-csi-driver/pkg/blobfuse.driverVersion=latest}" -o _output/blobfuseplugin ./pkg/blobfuseplugin
+RUN CGO_ENABLED=0 GOOS=$TARGETOS GOARCH=$TARGETARCH go build -a -ldflags "${LDFLAGS:--X sigs.k8s.io/blobfuse-csi-driver/pkg/blobfuse.driverVersion=latest}" -o _output/blobplugin ./pkg/blobplugin
 
 FROM us.gcr.io/k8s-artifacts-prod/build-image/debian-base-amd64:v2.1.0
-COPY --from=builder /go/src/sigs.k8s.io/blobfuse-csi-driver/_output/blobfuseplugin /blobfuseplugin
+COPY --from=builder /go/src/sigs.k8s.io/blobfuse-csi-driver/_output/blobplugin /blobplugin
 RUN apt-get update && clean-install ca-certificates pkg-config libfuse-dev cmake libcurl4-gnutls-dev libgnutls28-dev uuid-dev libgcrypt20-dev wget
 RUN wget -O /tmp/packages-microsoft-prod.deb https://packages.microsoft.com/config/ubuntu/16.04/packages-microsoft-prod.deb && dpkg -i /tmp/packages-microsoft-prod.deb && apt-get update && apt install blobfuse fuse -y && rm -f /tmp/packages-microsoft-prod.deb
 RUN apt remove wget -y
@@ -30,4 +30,4 @@ RUN apt install nfs-common -y || true
 LABEL maintainers="andyzhangx"
 LABEL description="Azure Blob Storage CSI driver"
 
-ENTRYPOINT ["/blobfuseplugin"]
+ENTRYPOINT ["/blobplugin"]
