+++ test/e2e/testsuites/dynamically_provisioned_reclaim_policy_tester.go
@@ -17,7 +17,7 @@ limitations under the License.
 package testsuites
 
 import (
-	"sigs.k8s.io/blobfuse-csi-driver/pkg/blobfuse"
+	"sigs.k8s.io/blobfuse-csi-driver/pkg/blob"
 	"sigs.k8s.io/blobfuse-csi-driver/test/e2e/driver"
 
 	v1 "k8s.io/api/core/v1"
